[BE] fix: 투표 기능 에러 수정

See merge request s08-bigdata-recom-sub2/S08P22C201!126

+++ back/tink/src/main/java/com/ssafy/tink/db/entity/DifficultyVoteId.java
@@ -3,7 +3,7 @@ package com.ssafy.tink.db.entity;
 import java.io.Serializable;
 
 public class DifficultyVoteId implements Serializable {
-	private int member;
+	private Long member;
 
 	private int pattern;
 }
